requirements.txt
pydantic
openai
anthropic
google-genai
python-dotenv
